refactor! : refactor of the evidence package and introduce utils.py
* remove of config.yaml, not used

* MOVE from evidence.py to_compact_text, to_totto_text,
clean_content to evidence.utils.py

* add logger.py
* add TableException and TableExceptionType
* add possibility to check which discarded page
