+++ src/claim/claim_generator.py
@@ -1,23 +1,26 @@
 from abc import abstractmethod
+
+from logger import logger
 from ..pipeline import PipelineElement
 from .claim import TextualClaim
 
 
 class TextualClaimGenerator(PipelineElement):
-    '''
+    """
     Element generating textual claims starting from Evidence objects.
-    '''
-    def __init__(self,encoding):
+    """
+
+    def __init__(self, encoding):
         self.encoding = encoding
 
     def generate(self, evidence, verbose=False):
-        '''
+        """
         Generate textual claims based on evidence
 
         :param evidence: a list of evidence objects
         :param verbose: if True prints additional debug informations
         :return: a list of TextualClaim objects
-        '''
+        """
         claims = []
         for i, e in enumerate(evidence):
             evidence_text = e.to_text(self.encoding)
@@ -25,33 +28,35 @@ class TextualClaimGenerator(PipelineElement):
             claim_json = TextualClaimGenerator._evidence_to_json(i, e, claim_text)
             claim = TextualClaim(claim_text, e, claim_json)
             claims.append(claim)
-            if verbose: print(claim)
+            if verbose:
+                logger.info(claim)
         return claims
 
     @abstractmethod
     def _generate_claim(self, text):
-        '''
-        Given an encoded piece of Evidence generates the corresponding claim in textual form.
+        """
+        Given an encoded piece of Evidence generates
+        the corresponding claim in textual form.
         Must be implemented in child class.
 
         :param text: encoded piece of Evidence
         :return: a claim in textual form
-        '''
+        """
         raise NotImplementedError("Must have implemented this.")
 
-    #TODO: move this in claim.py
+    # TODO: move this in claim.py
     @staticmethod
     def _evidence_to_json(sample_id,
                           evidence,
                           claim):
-        '''
+        """
         Encodes an Evidence object and its corresponding TextualClaim in JSON format
 
         :param sample_id:
         :param evidence:
         :param claim:
         :return: a dict object ready to be serialized
-        '''
+        """
         content = []
         context = {}
         for piece in evidence.evidence_pieces:
