+++ src/claim/totto_generator.py
@@ -13,7 +13,8 @@ class ToTToGenerator(TextualClaimGenerator):
         self.model = tf.saved_model.load(model_path)
         self.model_func = self.model.signatures[tf.saved_model.DEFAULT_SERVING_SIGNATURE_DEF_KEY]
         if encoding != 'totto' and encoding != 'compact':
-            raise ValueError("Invalid encoding choice.")
+            raise ValueError("Invalid encoding choice. Expected 'totto' or 'compact'"
+                             f"received {encoding}")
         self.encoding = encoding
 
     def _generate_claim(self, text):
