+++ src/evidence/evidence_retriever.py
@@ -1,6 +1,6 @@
 from abc import abstractmethod
 
-from ..pipeline import PipelineElement
+from pipeline import PipelineElement
 
 
 class EvidenceRetriever(PipelineElement):
