+++ src/evidence/feverous_retriever/__init__.py
@@ -1,0 +1,7 @@
+from .feverous_retriever_random import FeverousRetrieverRandom
+from .feverous_retriever_template import FeverousTemplateRetriever
+
+__all__ = [
+    FeverousRetrieverRandom,
+    FeverousTemplateRetriever
+]
