+++ src/evidence/feverous_retriever/feverous_retriever_random/__init__.py
@@ -1,0 +1,6 @@
+from .feverous_retriever_random import FeverousRetrieverRandom
+
+__all__ = [
+    FeverousRetrieverRandom,
+
+]
