+++ src/evidence/feverous_retriever/feverous_retriever_template/__init__.py
@@ -1,0 +1,5 @@
+from .feverous_retriever_template import FeverousTemplateRetriever
+
+__all__ = [
+    FeverousTemplateRetriever
+]
