+++ src/evidence/feverous_retriever/feverous_retriever_template/feverous_retriever_template.py
@@ -1,0 +1,2 @@
+class FeverousTemplateRetriever:
+    pass
