+++ src/pipeline/pipeline.py
@@ -22,14 +22,15 @@ class ClaimGeneratorPipeline:
         self.elements = elements
 
     @abstractmethod
-    def generate(self, table):
+    def generate(self, table, header_content=False):
         """
-        Runs the whole pipeline on the provided table
+           Runs the whole pipeline on the provided table
 
-        :param self
-        :param table: table to retrieve textual claims from.
-        :returns A list of textual claims
-        """
+           :param self
+           :param table: table to retrieve textual claims from.
+           :param header_content: if True, return the header inside the content
+           :returns A list of textual claims
+       """
         next_input = table
         # Calls __call__ method for each element in sequence
         for e in self.elements:
